+++ packages/sdk/src/pub/types.ts
@@ -69,7 +69,12 @@ export {
     TypedJsonParseErrorCode,
     TypedJson,
 } from '../types/util.js';
-export { jsonParse, jsonStringify } from '../types/json.js';
+export {
+    jsonParse,
+    jsonStringify,
+    jsonUnwrapStringify,
+    BigintFormatType,
+} from '../types/json.js';
 
 // These cannot be exported directly as modules because of a bug in an eslint plugin.
 // https://github.com/import-js/eslint-plugin-import/issues/2289.