+++ packages/sdk/src/types/AccountAddress.ts
@@ -8,6 +8,11 @@ import {
 } from './util.js';
 import { Base58String } from '../types.js';
 
+// IMPORTANT:
+// When adding functionality to this module, it is important to not change the wrapper class, as changing this might break compatibility
+// between different versions of the SDK, e.g. if a dependency exposes an API that depends on the class and a class from a different version
+// of the SDK is passed.
+
 /**
  * The {@linkcode TypedJsonDiscriminator} discriminator associated with {@linkcode Type} type.
  */
@@ -28,6 +33,16 @@ class AccountAddress {
         /** The account address represented in bytes. */
         public readonly decodedAddress: Uint8Array
     ) {}
+}
+
+/**
+ * Unwraps {@linkcode Type} value
+ *
+ * @param value value to unwrap.
+ * @returns the unwrapped {@linkcode Serializable} value
+ */
+export function toUnwrappedJSON(value: Type): Serializable {
+    return toBase58(value);
 }
 
 /**