+++ packages/sdk/src/types/ContractName.ts
@@ -6,6 +6,11 @@ import {
     makeFromTypedJson,
 } from './util.js';
 
+// IMPORTANT:
+// When adding functionality to this module, it is important to not change the wrapper class, as changing this might break compatibility
+// between different versions of the SDK, e.g. if a dependency exposes an API that depends on the class and a class from a different version
+// of the SDK is passed.
+
 /**
  * The {@linkcode TypedJsonDiscriminator} discriminator associated with {@linkcode Type} type.
  */
@@ -20,6 +25,16 @@ class ContractName {
         /** The internal string value of the contract name. */
         public readonly value: string
     ) {}
+}
+
+/**
+ * Unwraps {@linkcode Type} value
+ *
+ * @param value value to unwrap.
+ * @returns the unwrapped {@linkcode Serializable} value
+ */
+export function toUnwrappedJSON(value: Type): Serializable {
+    return toString(value);
 }
 
 /** The name of a smart contract. Note: This does _not_ including the 'init_' prefix. */