+++ packages/sdk/src/types/CredentialRegistrationId.ts
@@ -7,6 +7,11 @@ import {
     makeFromTypedJson,
 } from './util.js';
 
+// IMPORTANT:
+// When adding functionality to this module, it is important to not change the wrapper class, as changing this might break compatibility
+// between different versions of the SDK, e.g. if a dependency exposes an API that depends on the class and a class from a different version
+// of the SDK is passed.
+
 /**
  * The {@linkcode TypedJsonDiscriminator} discriminator associated with {@linkcode Type} type.
  */
@@ -28,7 +33,7 @@ class CredentialRegistrationId {
         public readonly credId: string
     ) {}
 
-    public toJSON(): string {
+    public toJSON(): Serializable {
         return this.credId;
     }
 }