+++ packages/sdk/src/types/EntrypointName.ts
@@ -5,6 +5,11 @@ import {
     makeFromTypedJson,
 } from './util.js';
 
+// IMPORTANT:
+// When adding functionality to this module, it is important to not change the wrapper class, as changing this might break compatibility
+// between different versions of the SDK, e.g. if a dependency exposes an API that depends on the class and a class from a different version
+// of the SDK is passed.
+
 /**
  * The {@linkcode TypedJsonDiscriminator} discriminator associated with {@linkcode Type} type.
  */
@@ -26,6 +31,16 @@ class EntrypointName<S extends string = string> {
         /** The internal string value of the receive name. */
         public readonly value: S
     ) {}
+}
+
+/**
+ * Unwraps {@linkcode Type} value
+ *
+ * @param value value to unwrap.
+ * @returns the unwrapped {@linkcode Serializable} value
+ */
+export function toUnwrappedJSON<S extends string>(value: Type<S>): S {
+    return toString(value);
 }
 
 /**