+++ packages/sdk/test/ci/types/json.test.ts
@@ -18,8 +18,10 @@ import {
     TransactionExpiry,
     ModuleReference,
     DataBlob,
+    jsonUnwrapStringify,
     jsonStringify,
     jsonParse,
+    BigintFormatType,
 } from '../../../src/pub/types.js';
 
 describe('JSON ID test', () => {
@@ -71,7 +73,7 @@ describe('JSON ID test', () => {
     });
 });
 
-describe('jsonStringify', () => {
+describe(jsonStringify, () => {
     test('Throws on circular reference', () => {
         const obj = {};
         // eslint-disable-next-line @typescript-eslint/no-explicit-any
@@ -83,5 +85,278 @@ describe('jsonStringify', () => {
     test('Allow non-circular references to same object', () => {
         const other = { test: 1 };
         expect(() => jsonStringify([other, other])).not.toThrow();
+    });
+});
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const testBigintReplacer = (_k: any, v: any) =>
+    typeof v === 'bigint' ? 'replaced' : v;
+
+// eslint-disable-next-line @typescript-eslint/no-explicit-any
+const unsafeReplacer = (_: any, v: any) =>
+    typeof v === 'bigint' ? Number(v) : v;
+
+describe(jsonUnwrapStringify, () => {
+    const t = 100n;
+
+    test('Serializes bigint values as expected', () => {
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual('100');
+        expect(jsonUnwrapStringify(t, BigintFormatType.String)).toEqual(
+            '"100"'
+        );
+        expect(
+            jsonUnwrapStringify(t, BigintFormatType.None, testBigintReplacer)
+        ).toEqual('"replaced"');
+
+        // Test for numbers bigger than `Number.MAX_SAFE_INTEGER`
+        const unsafeNumber = SequenceNumber.create(9007199254740997n);
+        const unsafeExpected = '9007199254740997';
+        expect(
+            jsonUnwrapStringify(unsafeNumber, undefined, unsafeReplacer)
+        ).not.toEqual(unsafeExpected);
+        expect(
+            jsonUnwrapStringify(unsafeNumber, BigintFormatType.Integer)
+        ).toEqual(unsafeExpected);
+    });
+
+    test('Throws `TypeError` on serialize bigint', () => {
+        expect(() => jsonUnwrapStringify(t)).toThrowError(TypeError);
+    });
+
+    test('Serializes nested bigint (arrays) values as expected', () => {
+        const t = [100n, 200n];
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(
+            '[100,200]'
+        );
+        expect(jsonUnwrapStringify(t, BigintFormatType.String)).toEqual(
+            '["100","200"]'
+        );
+        expect(
+            jsonUnwrapStringify(t, BigintFormatType.None, testBigintReplacer)
+        ).toEqual('["replaced","replaced"]');
+    });
+
+    test('Serializes nested bigint (objects) values as expected', () => {
+        const t = { a: 100n, b: 200n };
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(
+            '{"a":100,"b":200}'
+        );
+        expect(jsonUnwrapStringify(t, BigintFormatType.String)).toEqual(
+            '{"a":"100","b":"200"}'
+        );
+        expect(
+            jsonUnwrapStringify(t, BigintFormatType.None, testBigintReplacer)
+        ).toEqual('{"a":"replaced","b":"replaced"}');
+    });
+
+    test('Replacer overrides bigint default', () => {
+        expect(
+            jsonUnwrapStringify(t, BigintFormatType.Integer, testBigintReplacer)
+        ).toEqual('"replaced"');
+        expect(
+            jsonUnwrapStringify(t, BigintFormatType.String, testBigintReplacer)
+        ).toEqual('"replaced"');
+    });
+});
+
+describe('ContractName', () => {
+    test('Unwraps as expected', () => {
+        const t = ContractName.fromString('some-name');
+        const e = '"some-name"';
+
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('InitName', () => {
+    test('Unwraps as expected', () => {
+        const t = InitName.fromString('init_some-name');
+        const e = '"init_some-name"';
+
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('ReceiveName', () => {
+    test('Unwraps as expected', () => {
+        const t = ReceiveName.fromString('some_name.test');
+        const e = '"some_name.test"';
+
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('EntrypointName', () => {
+    test('Unwraps as expected', () => {
+        const t = EntrypointName.fromString('some_name.test');
+        const e = '"some_name.test"';
+
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('TransactionExpiry', () => {
+    test('Unwraps as expected', () => {
+        const t = TransactionExpiry.fromEpochSeconds(300);
+        const e = '300';
+
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('CcdAmount', () => {
+    test('Unwraps as expected', () => {
+        let t = CcdAmount.fromMicroCcd(300);
+        let e = '"300"';
+
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+
+        // Test for numbers bigger than Number.MAX_SAFE_INTEGER
+        t = CcdAmount.fromMicroCcd(9007199254740997n);
+        e = '"9007199254740997"';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('SequenceNumber', () => {
+    test('Unwraps as expected', () => {
+        const t = SequenceNumber.create(300);
+        const e = '300';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('Energy', () => {
+    test('Unwraps as expected', () => {
+        let t = Energy.create(300);
+        let e = '300';
+
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+
+        // Test for numbers bigger than Number.MAX_SAFE_INTEGER
+        t = Energy.create(9007199254740997n);
+        e = '9007199254740997';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('Timestamp', () => {
+    test('Unwraps as expected', () => {
+        let t = Timestamp.fromMillis(300);
+        let e = '300';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+
+        // Test for numbers bigger than Number.MAX_SAFE_INTEGER
+        t = Timestamp.fromMillis(9007199254740997n);
+        e = '9007199254740997';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('Duration', () => {
+    test('Unwraps as expected', () => {
+        let t = Duration.fromMillis(300);
+        let e = '300';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+
+        // Test for numbers bigger than Number.MAX_SAFE_INTEGER
+        t = Duration.fromMillis(9007199254740997n);
+        e = '9007199254740997';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('ContractAddress', () => {
+    test('Unwraps as expected', () => {
+        let t = ContractAddress.create(100, 10);
+        let e = '{"index":100,"subindex":10}';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+        expect(jsonUnwrapStringify(t, BigintFormatType.String)).toEqual(
+            '{"index":"100","subindex":"10"}'
+        );
+
+        // Test for numbers bigger than Number.MAX_SAFE_INTEGER
+        t = ContractAddress.create(9007199254740997n, 10);
+        e = '{"index":9007199254740997,"subindex":10}';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+
+        t = ContractAddress.create(9007199254740997n, 109007199254740997n);
+        e = '{"index":9007199254740997,"subindex":109007199254740997}';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('Parameter', () => {
+    test('Unwraps as expected', () => {
+        const t = Parameter.fromHexString('000102');
+        const e = '"000102"';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('TransactionHash', () => {
+    test('Unwraps as expected', () => {
+        const v =
+            '1a17008f7944a5fd11a665a864266fb2d76794e754986c367455a4937fd3a66b';
+        const t = TransactionHash.fromHexString(v);
+        const e = `"${v}"`;
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('BlockHash', () => {
+    test('Unwraps as expected', () => {
+        const v =
+            '1a17008f7944a5fd11a665a864266fb2d76794e754986c367455a4937fd3a66b';
+        const t = BlockHash.fromHexString(v);
+        const e = `"${v}"`;
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('ReturnValue', () => {
+    test('Unwraps as expected', () => {
+        const t = ReturnValue.fromHexString('000102');
+        const e = '"000102"';
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('ModuleReference', () => {
+    test('Unwraps as expected', () => {
+        const v =
+            '5d99b6dfa7ba9dc0cac8626754985500d51d6d06829210748b3fd24fa30cde4a';
+        const t = ModuleReference.fromHexString(v);
+        const e = `"00000020${v}"`; // is prefixed with 4 byte length
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('CredentialRegistrationId', () => {
+    test('Unwraps as expected', () => {
+        const v =
+            '83e4b29e1e2582a6f1dcc93bf2610ce6b0a6ba89c8f03e661f403b4c2e055d3adb80d071c2723530926bb8aed3ed52b1';
+        const t = CredentialRegistrationId.fromHexString(v);
+        const e = `"${v}"`;
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('DataBlob', () => {
+    test('Unwraps as expected', () => {
+        const v =
+            '83e4b29e1e2582a6f1dcc93bf2610ce6b0a6ba89c8f03e661f403b4c2e055d3adb80d071c2723530926bb8aed3ed52b1';
+        const t = new DataBlob(Buffer.from(v, 'hex'));
+        const e = `"0030${v}"`; // Is prefixed with 2 byte length
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
+    });
+});
+
+describe('AccountAddress', () => {
+    test('Unwraps as expected', () => {
+        const v = '4owvMHZSKsPW8QGYUEWSdgqxfoPBh3ZwPameBV46pSvmeHDkEe';
+        const t = AccountAddress.fromBase58(v);
+        const e = `"${v}"`; // Is prefixed with 2 byte length
+        expect(jsonUnwrapStringify(t, BigintFormatType.Integer)).toEqual(e);
     });
 });